fix: Contexts on server not working.

+++ server/server.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"encoding/json"
+	"errors"
 	"fmt"
 	_ "github.com/mattn/go-sqlite3"
 	"io"
@@ -21,11 +22,13 @@ type Quotation struct {
 func init() {
 	os.MkdirAll("../db", 0700)
 
-	file, err := os.Create("../db/database.db")
-	if err != nil {
-		fmt.Println("ERROR:", err)
+	if _, err := os.Stat("../db/database.db"); errors.Is(err, os.ErrNotExist) {
+		file, err := os.Create("../db/database.db")
+		if err != nil {
+			fmt.Println("ERROR:", err)
+		}
+		file.Close()
 	}
-	file.Close()
 
 	db, err := sql.Open("sqlite3", "file:../db/database.db")
 	if err != nil {
@@ -40,9 +43,6 @@ func init() {
 }
 
 func main() {
-
-	// Two context: For DB: 10ms, For API: 200ms
-
 	// Endpoint available to client: /cotacao on port 8080
 	http.HandleFunc("/cotacao", GetDollarQuotationHandler)
 	fmt.Println("Running server on port:8080")
@@ -53,77 +53,82 @@ func main() {
 // GetDollarQuotationHandler Call API https://economia.awesomeapi.com.br/json/last/USD-BRL and return JSON to client
 func GetDollarQuotationHandler(w http.ResponseWriter, r *http.Request) {
 
-	ctx := r.Context()
+	// Two context: For DB: 10ms, For API: 200ms
+	ctxRequest, cancel := context.WithTimeout(context.Background(), 200*time.Millisecond)
 	fmt.Println("Request started.")
+	defer cancel()
 	defer fmt.Println("Request ended.")
 
-	select {
-	case <-time.After(200 * time.Millisecond):
-		req, err := http.Get("https://economia.awesomeapi.com.br/json/last/USD-BRL")
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		defer req.Body.Close()
-
-		res, err := io.ReadAll(req.Body)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		var quotation Quotation
-		err = json.Unmarshal(res, &quotation)
-		if err != nil {
-			fmt.Println("ERROR:", err)
-			return
-		}
-
-		SaveQuotationOnDatabase(quotation)
-
-		fmt.Println("Request processed.")
-		json.NewEncoder(w).Encode(quotation)
-		return
-	case <-ctx.Done():
-		fmt.Println("Request cancelled")
-		http.Error(w, "Request cancelled by client.", http.StatusRequestTimeout)
+	req, err := http.NewRequestWithContext(ctxRequest, "GET", "https://economia.awesomeapi.com.br/json/last/USD-BRL", nil)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
+	res, err := http.DefaultClient.Do(req)
+
+	if err != nil {
+		if errors.Is(err, context.DeadlineExceeded) {
+			fmt.Println("Request timeout reached.")
+			w.WriteHeader(http.StatusRequestTimeout)
+			return
+		}
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	defer res.Body.Close()
+
+	resp, err := io.ReadAll(res.Body)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	var quotation Quotation
+	err = json.Unmarshal(resp, &quotation)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	err = SaveQuotationOnDatabase(quotation)
+	if err != nil {
+		if errors.Is(err, context.DeadlineExceeded) {
+			fmt.Println("Save database timeout reached.")
+			w.WriteHeader(http.StatusRequestTimeout)
+			return
+		}
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	json.NewEncoder(w).Encode(quotation)
 
 }
 
 // SaveQuotationOnDatabase Save information on SQLite
 func SaveQuotationOnDatabase(quotation Quotation) error {
 
-	dbCtx := context.Background()
+	dbCtx, cancel := context.WithTimeout(context.Background(), 10*time.Millisecond)
 	fmt.Println("Save database started.")
-	dbCtx, cancel := context.WithTimeout(dbCtx, 10*time.Millisecond)
 	defer cancel()
 	defer fmt.Println("Save database ended.")
 
-	select {
-	case <-dbCtx.Done():
-		fmt.Println("Save database cancelled.")
-		fmt.Println(dbCtx.Err())
-		return dbCtx.Err()
-	case <-time.After(10 * time.Millisecond):
-		db, err := sql.Open("sqlite3", "file:../db/database.db")
-		if err != nil {
-			fmt.Println("ERROR:", err)
-		}
-		defer db.Close()
+	db, err := sql.Open("sqlite3", "file:../db/database.db")
+	if err != nil {
+		return err
+	}
+	defer db.Close()
 
-		stmt, err := db.Prepare("INSERT INTO quotation(bid) values(?)")
-		if err != nil {
-			fmt.Println("ERROR:", err)
-		}
+	stmt, err := db.PrepareContext(dbCtx, "INSERT INTO quotation(bid) values(?)")
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
 
-		_, err = stmt.Exec(quotation.Usdbrl.Bid)
-		if err != nil {
-			fmt.Println("ERROR:", err)
-		}
-		fmt.Println("Save database sucessfully")
-		return nil
+	_, err = stmt.ExecContext(dbCtx, quotation.Usdbrl.Bid)
+	if err != nil {
+		return err
 	}
 
+	return nil
 }
